Ocr improvements (#5)
Changed ocr to use config= --psm 12 which has increased accuracy at the cost of reading line formatting correctly. This also enables reading the full image in one go allowing labbie to be flexible to handle different types of labs, and numbers of lab enchants. To counter act the introduced downside the read enchants from the ocr are compared against the complete list of enchants using a trie for prefix matching for efficiency.

Other improvement to the OCR that was introduced is to do the grayscale over all of the provided color channels.
CV2 was also added to the project, in order to use different morphologies to improve accuracy. First one added is a Dilation pass to thicken the characters. This is turned off unless you modify the constants file.