0.6.0 (#6)
* Assorted changes

- Added compression for enchant data to reduce bandwidth requirements
- Adjusted the About window
  - Minor UI changes
  - Addition of scrape dates for reference
- Reduced logspam a bit in debug mode
- Added auto-refresh for enchant data, so Labbie will attempt to download fresh data once the Lab rolls over to the next day

* Reduce logspam

* Bugfix

* Center windows

* Progress on price checking and new results UI

* WIP

* Working, just need improved OCR now

* UI updates

* Fix base searches

* Use new mod format

* Fixup mods

* Removed default

* Ocr improvements (#5)

Changed ocr to use config= --psm 12 which has increased accuracy at the cost of reading line formatting correctly. This also enables reading the full image in one go allowing labbie to be flexible to handle different types of labs, and numbers of lab enchants. To counter act the introduced downside the read enchants from the ocr are compared against the complete list of enchants using a trie for prefix matching for efficiency.

Other improvement to the OCR that was introduced is to do the grayscale over all of the provided color channels.
CV2 was also added to the project, in order to use different morphologies to improve accuracy. First one added is a Dilation pass to thicken the characters. This is turned off unless you modify the constants file.

* Final changes

* Switch tab on search

* Change OCR matching algo

* OCR adjustment

* Unkrangle OCR

* Added location save (auto) and reset (manual)

* Bugfixes related to OCR and hotkeys

* UI changes

- Added hint to prompt right click on search results
- Hid search window from alt tab / taskbar
- Allowed OCR hotkey to open the search window any time

* Allow OCR hotkey to toggle window

Co-authored-by: Caithir <colter.norick@gmail.com>